Grandes mudanças no front, proteção na deleção de users e prester a implentar menu admin

+++ backend/controllers/deleteUser.js
@@ -1,18 +1,23 @@
 import {deleteUser, findUserByEmail} from "../models/userModel.js"
-
 export async function deletarUsuario(req,res){
-
     const {email} = req.body
     const userExists = await findUserByEmail(email)
-    if (!userExists){
-        return res.status(404).json({ message: "Email Nao encontrado" })
-    }
-    try{
-        await deleteUser(email)
-        res.json({message: "Usuário Deletado"})
-    }
-    catch{
-        res.status(500).json({ error: "Erro ao deletar usuário" })
+    if(req.user.role == 'admin'){
+
+        if (!userExists){
+            return res.status(404).json({ message: "Email Nao encontrado" })
+        }
+        try{
+                await deleteUser(email)
+                res.json({message: "Usuário Deletado",})
+            }   
+
+        catch{
+            res.status(500).json({ error: "Erro ao deletar usuário" })
+        }
+    
+}
+    else{
+        res.status(403).json({ error: "Você não tem permissão pra executar esta ação" })
     }
 }
-
